SLinear and SLSTM in separated files
