+++ .vscode/settings.json
@@ -224,5 +224,6 @@
         "__string": "cpp",
         "filesystem": "cpp",
         "execution": "cpp"
-    }
+    },
+    "C_Cpp.errorSquiggles": "disabled"
 }
