+++ include/bindings/main_bindings.h
@@ -26,4 +26,6 @@
 #include "pooling_layer_bindings.h"
 #include "resnet_block_bindings.h"
 #include "sequential_bindings.h"
+#include "slinear_layer_bindings.h"
+#include "slstm_layer_bindings.h"
 #include "utils_bindings.h"
