+++ include/bindings/slstm_layer_bindings.h
@@ -11,6 +11,6 @@
 #pragma once
 #include <pybind11/pybind11.h>
 
-#include "../lstm_layer.h"
+#include "../slstm_layer.h"
 
-void bind_lstm_layer(pybind11::module_& modo);
+void bind_slstm_layer(pybind11::module_& modo);
