+++ include/sequential.h
@@ -21,6 +21,8 @@
 #include "linear_layer.h"
 #include "lstm_layer.h"
 #include "output_layer_update_cpu.h"
+#include "slinear_layer.h"
+#include "slstm_layer.h"
 #ifdef USE_CUDA
 #include "data_struct_cuda.cuh"
 #endif
